修改sequence生成方式 修复一些问题

+++ src/main/java/sunnn/sunsite/dto/request/DeletePicture.java
@@ -11,6 +11,6 @@ import javax.validation.constraints.NotBlank;
 @ToString
 public class DeletePicture {
 
-    private int pictureSequence;
+    private long pictureSequence;
 
 }
